use uint32 typed key for v6 reverse nat map

+++ pkg/service/service.go
@@ -664,7 +664,7 @@ func (s *Service) populateReverseNatMapV2FromV1(ipv4, ipv6 bool) error {
 				newKey = lbmap.NewRevNat4KeyV2(uint32(origKey.GetKey()))
 			} else {
 				origKey := key.(*lbmap.RevNat6Key)
-				newKey = lbmap.NewRevNat6KeyV2(uint16(origKey.GetKey()))
+				newKey = lbmap.NewRevNat6KeyV2(uint32(origKey.GetKey()))
 			}
 
 			err := v2Map.Update(newKey, value)
